adds base computer vision applications for identifying and tracking the corn
This commit adds two computer vision files. The first is a calibration file meant to be run
on the local computer. This file, the VisionCalibration.py, is used to isolate the HSV
color values needed to identify a variety of different colors and then saves those values
in a database. The next application, the VisionApplication, takes the HSV values from the
database and uses them to track multiple colors in one video stream. These files
are not yet configured to run on the Docker container, nor on the ros OS.