edits visionApplication to allow for multiple cameras
This commit changes the vision application to allow for multiple
cameras that process different colors. This will allow the robot
to be able to look for the base of the corn with one camera to
position the robot and the camera the other camera will be able to
be used to identify the corn. this code still needs to be debugged
given that it seems like the program keeps identifiying contours
at the last position they were seen even after they were taken
from the view of the camera.